Add models folder and place models inside
