+++ client/models/Favourite.ts
@@ -1,0 +1,6 @@
+import { Repository } from "./Repository";
+
+export interface Favourite {
+    userId: string;
+    repositories: Repository[];
+  }
