+++ client/models/Repository.ts
@@ -1,0 +1,4 @@
+export interface Repository {
+    repoId: string;
+  }
+
